Refactor RubinTV services completely.
Now have a single top level automatic isr runner, and TV channel
services which sit below waiting for their data products to arrive.

This refactoring will allow Gen3 migration to be much easier, as well
as removing duplicate processing, and allowing for easy creation of
summit daemons to be run on the newly provisioned dm-bastion machine
at the summit.